Manually merging changes made in omnisia-recursia-rrt-mml-2019 repo.
Modified COSIATECEncoding.java so that it uses TECQualityComparator that takes a comparator string as an argument
that determinest the order of priority in which the heuristics are applied when determining whether one TEC is
better than another (lines 195-197 and 229-237).

In DrawPoints.java, removed calls to noLoop in lines 320, 326, 350 and 355 along with calls to exit in lines 351 and 356 and added a general
exit in line 358 in the draw method so that the program terminates properly after drawing an image.

In OMNISIA.java, added the DRAW_POINT_SET switch (lines 97, 139, 244-246, 614, 729-732, 794, 805-832, 866, 1125),
the RHYTHM_ONLY switch (98, 140, 225-227, 615, 733-736, 1141), the TEC_PRIORITY_STRING switch and the DUAL_TEC_PRIORITY_STRING switch
(99-100, 141-142, 156-160, 162-168, 616-617,737-752, 910-911, 1142-43).

In Encoding.java, in drawOccurrenceSetsToFile, stored runnable object in a variable and passed to invokeAndWait instead of constructing it on the fly.
Added changes for RHYTHM_ONLY switch (170-175) that effectively projects the dataset onto the x-axis by making all the y-values 0. Perhaps could be better to give each pattern
a different y-value instead (see https://github.com/chromamorph/omnisia/issues/2#issue-869781010).

In TECQualityComparator.java added support for priority string. Added priorityString variable (45) and a static DEFAULT_PRIORITY_STRING (46).
Added a constructor that takes a priority string as an argument (lines 80-83). Modified the compare method (lines 94-155) so that it applies
the heuristics in the order given in the priorityString. The former compare method is kept in the compareOld method (lines 157-197).